Rename auth fragments module
