+++ src/client/screens/ProjectScreen.tsx
@@ -20,7 +20,6 @@ import { resetProjectState } from "../actions/root";
 import { userFetch } from "../actions/user";
 import "../App.css";
 import AdvancedEditingModal from "../components/map/AdvancedEditingModal";
-import AuthModal from "../components/AuthModal";
 import CenteredContent from "../components/CenteredContent";
 import CopyMapModal from "../components/CopyMapModal";
 import Map from "../components/map/Map";
@@ -172,7 +171,6 @@ const ProjectScreen = ({
                   geoLevels={staticMetadata.geoLevelHierarchy}
                 />
               )}
-              <AuthModal project={project} />
               <CopyMapModal project={project} />
               <Flex id="tour-start" sx={style.tourStart}></Flex>
             </React.Fragment>
